+++ needle/core/framework/framework.py
@@ -115,14 +115,6 @@ class Framework(cmd.Cmd):
     # ==================================================================================================================
     # COMPLETE METHODS
     # ==================================================================================================================
-    def complete_keys(self, text, line, *ignored):
-        args = line.split()
-        options = ['list', 'add', 'delete']
-        if 1 < len(args) < 4:
-            if args[1].lower() in options[:1]:
-                return []
-        return [x for x in options if x.startswith(text)]
-
     def complete_load(self, text, *ignored):
         return [x for x in Framework._loaded_modules if x.startswith(text)]
     complete_use = complete_load
@@ -347,12 +339,28 @@ class Framework(cmd.Cmd):
         print(getattr(self, 'do_jobs').__doc__)
         print('')
         print('Usage: jobs')
+        print('...list background jobs currently running.')
         print('')
 
     def help_kill(self):
         print(getattr(self, 'do_kill').__doc__)
         print('')
-        print('Usage: <job number>')
+        print('Usage: kill <job number>')
+        print('...stop the background job specified.')
+        print('')
+
+    def help_issues(self):
+        print(getattr(self, 'do_issues').__doc__)
+        print('')
+        print('Usage: issues')
+        print('...list the issues already identified.')
+        print('')
+
+    def help_add_issue(self):
+        print(getattr(self, 'do_add_issue').__doc__)
+        print('')
+        print('Usage: add_issue')
+        print('...start a wizard that will allow to manually add an issue.')
         print('')
 
     # ==================================================================================================================
@@ -589,6 +597,14 @@ class Framework(cmd.Cmd):
             self.printer.error("Error while killing job: no job attached to the selected job number. Continuing...")
         except Exception:
             self.print_exception()
+
+    def do_issues(self, params):
+        """List currently gathered issues."""
+        self.ISSUE_MANAGER.issue_print()
+
+    def do_add_issue(self, params):
+        """Prompt the user to manually add an issue."""
+        self.ISSUE_MANAGER.issue_add_manual()
 
     # ==================================================================================================================
     # CONNECTION METHODS